# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HCIS YSQ API
  version: 0.0.1
  description: |
    Kontrak awal untuk vertical slice identity, employee profile, dan leave.
    Status dokumen DRAFT; endpoint belum menyatakan implementasi telah tersedia.
servers:
  - url: https://api.example.invalid
    description: Placeholder only
...
paths:
...
  /admin/leave/configuration:
    get:
      tags: [Leave]
      operationId: getAdminLeaveConfiguration
      summary: Read Leave policy configuration and the Organization-structure approval source
      responses:
        '200':
          description: Leave administration configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLeaveConfiguration'
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/leave/employees/{employeeId}/entitlement-group:
    patch:
      tags: [Leave]
      operationId: updateLeaveEntitlementGroup
      summary: Update explicit Leave policy classification independently of approval routing
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [group]
              properties:
                group:
                  type: [string, 'null']
                  enum: [education, non_education, null]
      responses:
        '200': { description: Entitlement group updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/leave/employees/{employeeId}/preview:
    get:
      tags: [Leave]
      operationId: previewAdminLeaveApproval
      summary: Preview Leave policy and approval through the submission authority service
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: date
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Approval preview resolved from the effective published Organization structure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLeaveApprovalPreview'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
...
  responses:
    Unauthenticated:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Actor is not authorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found or intentionally concealed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Request validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
  schemas:
...
    LeaveAuthorityStep:
      type: object
      required: [employeeId, sources]
      properties:
        employeeId: { type: string, format: uuid }
        sources:
          type: array
          minItems: 1
          items:
            type: string
            enum: [DIRECT_MANAGER, UNIT_APPROVER, GOVERNANCE_APPROVER]
    AdminLeaveConfiguration:
      type: object
      required: [policies, employees, summary, approvalSource]
      properties:
        policies:
          type: array
          items: { type: object, additionalProperties: true }
        employees:
          type: array
          items:
            type: object
            required: [id, fullName, employeeNumber, status]
            properties:
              id: { type: string, format: uuid }
              fullName: { type: string }
              employeeNumber: { type: string }
              status: { type: string, enum: [active, inactive, resigned] }
              unitId: { type: [string, 'null'], format: uuid }
              unitName: { type: [string, 'null'] }
              positionName: { type: [string, 'null'] }
              leaveEntitlementGroup: { type: [string, 'null'], enum: [education, non_education, null] }
        summary:
          type: object
          required: [activeEmployees, entitlementGroupConfigured]
          properties:
            activeEmployees: { type: integer, minimum: 0 }
            entitlementGroupConfigured: { type: integer, minimum: 0 }
        approvalSource: { type: string, const: organization_structure }
    AdminLeaveApprovalPreview:
      type: object
      required: [employee, referenceDate, approvalChain, approvalSource, annualLeave, warnings]
      properties:
        employee: { type: object, additionalProperties: true }
        referenceDate: { type: string, format: date }
        approvalChain:
          type: array
          items: { $ref: '#/components/schemas/LeaveAuthorityStep' }
        approvalSource: { type: string, const: organization_structure }
        annualLeave:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: true
        warnings:
          type: array
          items:
            type: object
            required: [code, message]
            properties:
              code: { type: string }
              message: { type: string }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        correlationId:
          type: string
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string